Refactored folder structures to pave the way for a merge.
