+++ algorithm_distillation/models/__init__.py
@@ -1,3 +1,3 @@
 from .gpt2 import GPT2AD
 
-__all__ = ["util", "GPT2AD"]
+__all__ = ["sb3_util", "util", "GPT2AD"]
