+++ algorithm_distillation/rl_tasks/ad.py
@@ -6,8 +6,8 @@ import torch
 from tqdm import tqdm
 
 from algorithm_distillation.models.ad_transformer import ADTransformer
+from algorithm_distillation.models.util import get_sequence
 
-from .models.util import get_sequence
 from .task import GymTask
 from .task_manager import TaskManager
