+++ algorithm_distillation/rl_tasks/task.py
@@ -8,7 +8,7 @@ import numpy as np
 import stable_baselines3
 from stable_baselines3.common.buffers import ReplayBuffer
 
-from algorithm_distillation.sb3_util.callback import RolloutCallback
+from algorithm_distillation.models.sb3_util.callback import RolloutCallback
 
 
 class Task(abc.ABC):
